requirements.txt
streamlit==1.28.0
transformers==4.35.0
torch==2.1.0
SpeechRecognition==3.10.0
PyAudio==0.2.14
geocoder==1.38.1
twilio==8.10.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
